requirements.txt
Flask
gunicorn
python-dotenv
Flask-WTF
